src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/importNode-unsupported-node-type-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: document.createElement(...).createShadowRoot is not a function
importNode should throw informative errors for unsupported node types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.importNode(doc) threw exception NotSupportedError: Failed to execute 'importNode' on 'Document': The node provided is a document, which may not be imported..
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
